requirements.txt
flask
numpy
opencv-python
torch
torchvision
tweepy
pandas
transformers
facenet-pytorch
fer
Werkzeug
gunicorn
tensorflow
moviepy==1.0.3
nltk
flair
textblob
scikit-learn
deepface
tf-keras
python-dotenv